Html Updates
Fixed some of the issues with my "col-".  Added the charts in there. Made some tweaks with the class names and trimmed some of the information for the site.

NOTE: I'm not sure if i'm supposed to put the google charts in it's own "section tag". This might be the cause of the lines not showing?